// app/src/generated/models/propertyAcquisitionResponseDepreciationMethod.ts
/**
 * Generated by orval v8.5.1 🍺
 * Do not edit manually.
 * Buurman API
 * Property Management API for Small Landlords
 * OpenAPI spec version: 0.1.0
 */

/**
 * Accounting method used for property depreciation
 * @nullable
 */
export type PropertyAcquisitionResponseDepreciationMethod =
  | (typeof PropertyAcquisitionResponseDepreciationMethod)[keyof typeof PropertyAcquisitionResponseDepreciationMethod]
  | null;

export const PropertyAcquisitionResponseDepreciationMethod = {
  STRAIGHT_LINE: 'STRAIGHT_LINE',
  DECLINING_BALANCE: 'DECLINING_BALANCE',
  NONE: 'NONE',
} as const;