// app/src/generated/models/propertyAcquisitionResponseAcquisitionType.ts
/**
 * Generated by orval v8.5.1 🍺
 * Do not edit manually.
 * Buurman API
 * Property Management API for Small Landlords
 * OpenAPI spec version: 0.1.0
 */

/**
 * How the property was acquired
 */
export type PropertyAcquisitionResponseAcquisitionType =
  (typeof PropertyAcquisitionResponseAcquisitionType)[keyof typeof PropertyAcquisitionResponseAcquisitionType];

export const PropertyAcquisitionResponseAcquisitionType = {
  PURCHASE: 'PURCHASE',
  INHERITANCE: 'INHERITANCE',
  GIFT: 'GIFT',
  FORECLOSURE: 'FORECLOSURE',
  AUCTION: 'AUCTION',
  OTHER: 'OTHER',
} as const;