// app/src/generated/models/propertyValuationResponseValuationType.ts
/**
 * Generated by orval v8.5.1 🍺
 * Do not edit manually.
 * Buurman API
 * Property Management API for Small Landlords
 * OpenAPI spec version: 0.1.0
 */

/**
 * Method or source of the property valuation
 */
export type PropertyValuationResponseValuationType =
  (typeof PropertyValuationResponseValuationType)[keyof typeof PropertyValuationResponseValuationType];

export const PropertyValuationResponseValuationType = {
  MARKET: 'MARKET',
  APPRAISAL: 'APPRAISAL',
  TAX_ASSESSED: 'TAX_ASSESSED',
  PURCHASE: 'PURCHASE',
  INSURANCE: 'INSURANCE',
  USER_ESTIMATE: 'USER_ESTIMATE',
} as const;