// app/src/generated/models/propertyFinancingResponseFinancingType.ts
/**
 * Generated by orval v8.5.1 🍺
 * Do not edit manually.
 * Buurman API
 * Property Management API for Small Landlords
 * OpenAPI spec version: 0.1.0
 */

/**
 * Type of financing instrument
 */
export type PropertyFinancingResponseFinancingType =
  (typeof PropertyFinancingResponseFinancingType)[keyof typeof PropertyFinancingResponseFinancingType];

export const PropertyFinancingResponseFinancingType = {
  MORTGAGE: 'MORTGAGE',
  LEASING: 'LEASING',
  LOAN: 'LOAN',
  LINE_OF_CREDIT: 'LINE_OF_CREDIT',
  PRIVATE_FINANCING: 'PRIVATE_FINANCING',
  OTHER: 'OTHER',
} as const;