// app/src/generated/models/propertyFinancingResponseRateType.ts
/**
 * Generated by orval v8.5.1 🍺
 * Do not edit manually.
 * Buurman API
 * Property Management API for Small Landlords
 * OpenAPI spec version: 0.1.0
 */

/**
 * Interest rate structure of the financing
 */
export type PropertyFinancingResponseRateType =
  (typeof PropertyFinancingResponseRateType)[keyof typeof PropertyFinancingResponseRateType];

export const PropertyFinancingResponseRateType = {
  FIXED: 'FIXED',
  VARIABLE: 'VARIABLE',
  INTEREST_ONLY: 'INTEREST_ONLY',
  HYBRID: 'HYBRID',
} as const;