// app/src/generated/models/propertyFinancingResponseStatus.ts
/**
 * Generated by orval v8.5.1 🍺
 * Do not edit manually.
 * Buurman API
 * Property Management API for Small Landlords
 * OpenAPI spec version: 0.1.0
 */

/**
 * Current status of the financing arrangement
 */
export type PropertyFinancingResponseStatus =
  (typeof PropertyFinancingResponseStatus)[keyof typeof PropertyFinancingResponseStatus];

export const PropertyFinancingResponseStatus = {
  ACTIVE: 'ACTIVE',
  PAID_OFF: 'PAID_OFF',
  REFINANCED: 'REFINANCED',
  DEFAULTED: 'DEFAULTED',
} as const;