// app/src/generated/models/financingPaymentResponseStatus.ts
/**
 * Generated by orval v8.5.1 🍺
 * Do not edit manually.
 * Buurman API
 * Property Management API for Small Landlords
 * OpenAPI spec version: 0.1.0
 */

/**
 * Status of a financing installment payment
 */
export type FinancingPaymentResponseStatus =
  (typeof FinancingPaymentResponseStatus)[keyof typeof FinancingPaymentResponseStatus];

export const FinancingPaymentResponseStatus = {
  SCHEDULED: 'SCHEDULED',
  COMPLETED: 'COMPLETED',
  MISSED: 'MISSED',
  LATE: 'LATE',
} as const;